refactor(atom-svelte): use Svelte-native naming, drop React terms
Address review feedback that React terminology was creeping into the adapter:

- Rename `useAtomSuspense` -> `useAtomPromise` (it returns a promise handle for
  `{#await}` / `<svelte:boundary>`; "Suspense" is a React concept).
- Rename `useAtomState` / `AtomState` -> `useAtomBindable` / `AtomBindable`
  (it mirrors a `$bindable` prop, not `$state`).
- Merge `Hooks.ts` into `index.ts` (like the Vue adapter) so there is no
  React-style "Hooks" module; rename the test file to `index.test.ts` and the
  async fixture to `Awaited.svelte`.
- Replace the `@category hooks` doc grouping with `@category bindings` and scrub
  "hooks" / "Suspense" wording from the module and `RegistryContext` docs.

The `useAtom*` function prefix is kept, matching the Vue and Solid adapters
(Vue composables use it too); no behaviour changes.
