+++ packages/atom/svelte/test/StateInput.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { setRegistry, useAtomState } from "@effect/atom-svelte"
+  import { setRegistry, useAtomBindable } from "@effect/atom-svelte"
   import type * as Atom from "effect/unstable/reactivity/Atom"
   import type * as AtomRegistry from "effect/unstable/reactivity/AtomRegistry"
 
@@ -12,8 +12,8 @@
   } = $props()
 
   setRegistry(registry)
-  const state = useAtomState(atom)
+  const bindable = useAtomBindable(atom)
 </script>
 
-<input data-testid="input" bind:value={state.current} />
-<span data-testid="value">{state.current}</span>
+<input data-testid="input" bind:value={bindable.current} />
+<span data-testid="value">{bindable.current}</span>
